Extract access level tests into separate test files
